feat: RedmineCLI.Commonライブラリへの共通コンポーネント移行
RedmineCLIとRedmineCLI.Extension.Boardで重複していたコンポーネントを
RedmineCLI.Commonライブラリに共通化しました。

## 共通化した主要コンポーネント：

### HTTPクライアント基盤
- RedmineHttpClientFactory: HTTPクライアント生成の統一
- Pollyリトライポリシー対応
- セッション認証とAPIキー認証の両方に対応

### 認証ヘルパー
- AuthenticationHelper: フォームベースログイン機能
- セッションクッキー管理
- 資格情報検証

### JSONシリアライゼーション
- DateTimeConverter: 複数の日時フォーマットサポート
- JsonSerializerOptionsProvider: 標準化されたJSONオプション

### 例外クラス
- RedmineApiException: API エラーハンドリング
- ValidationException: 入力検証エラー

## 後方互換性：
既存コードとの互換性を保つため、元のクラスは[Obsolete]属性付きで
残してあります。これにより段階的な移行が可能です。

## テスト結果：
- 全476件のテストが成功
- ビルドエラーなし
- コードフォーマット違反なし
