+++ RedmineCLI.Common/RedmineCLI.Common.csproj
@@ -18,6 +18,9 @@
   <ItemGroup>
     <!-- AOT互換のJSON処理 -->
     <PackageReference Include="System.Text.Json" Version="9.0.0" />
+    <!-- HTTPクライアント基盤 -->
+    <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.0" />
   </ItemGroup>
 
 </Project>
